refactor: set layout and root route to treat locales
