+++ website/src/app/[lang]/layout.tsx
@@ -3,11 +3,16 @@ import './layout.scss';
 
 import Footer from '@components/footer';
 import Header from '@components/header';
+import { AVAILABLE_LOCALES } from '@i18n/locales';
 
 export const metadata = {
   title: 'Welcome to website',
   description: 'Generated by create-nx-workspace',
 };
+
+export async function generateStaticParams() {
+  return AVAILABLE_LOCALES.map((lang) => ({ lang }));
+}
 
 export default function RootLayout({
   children,
