+++ website/tsconfig.json
@@ -23,6 +23,7 @@
       "@styles/*": ["website/src/styles/*"],
       "@config/*": ["website/src/config/*"],
       "@graphql/*": ["website/src/graphql/*"],
+      "@i18n/*": ["website/src/i18n/*"],
     },
   },
